Provide DAL.Sandbox helpers
This adds a module `DAL.Sandbox` with the helper
`checkout_all_as_shared` which, given a list of repos by module name
(atom) will check those repos out in `:shared` mode using `self()`. If
there are any unexpected values, we simply return `:error`, although it
might be nicer to make it the standard two-tuple as to provide a reason
for why the checkout has failed.